+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Support for SMS notifications for ONS
 - Support for publisher maps for service connector
 - Support to attach and detach instances to `oci_core_instance_pool` 
+- Support for GoldenGate service
 
 ## 4.16.0 (March 03, 2021)